requirements.txt
numpy
Flask
gunicorn
sklearn
pandas
PyYAML
evidently
dill
matplotlib
boto3
python-dotenv
-e .